services/user-service/app/security/jwt.py
from jose import jwt, JWTError
from app.core.config import settings
from typing import Optional, Dict, Any

def verify_jwt(token: str) -> Optional[Dict[str, Any]]:
    try:
        payload = jwt.decode(token, settings.secret_key, algorithms=[settings.jwt_algorithm])
        return payload
    except JWTError:
        return None